Move new itemlist classes into new package
