+++ examples/mini-examples/itemlist-example/src/main/java/de/saxsys/mvvmfx/examples/itemlist/ItemListExampleViewModel.java
@@ -3,8 +3,8 @@ package de.saxsys.mvvmfx.examples.itemlist;
 import de.saxsys.mvvmfx.ViewModel;
 import de.saxsys.mvvmfx.examples.itemlist.model.IceCreamFlavor;
 import de.saxsys.mvvmfx.examples.itemlist.model.IceCreamRepository;
-import de.saxsys.mvvmfx.utils.newitemlist.ItemList;
-import de.saxsys.mvvmfx.utils.newitemlist.ViewItemList;
+import de.saxsys.mvvmfx.utils.items.ItemList;
+import de.saxsys.mvvmfx.utils.items.ViewItemList;
 import javafx.beans.binding.Bindings;
 import javafx.beans.property.ReadOnlyBooleanProperty;
 import javafx.beans.property.ReadOnlyBooleanWrapper;
