+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/utils/newitemlist/ChoiceBoxItemListTest.java
@@ -2,6 +2,7 @@ package de.saxsys.mvvmfx.utils.newitemlist;
 
 import de.saxsys.mvvmfx.testingutils.FxTestingUtils;
 import de.saxsys.mvvmfx.testingutils.jfxrunner.JfxRunner;
+import de.saxsys.mvvmfx.utils.items.ItemList;
 import javafx.scene.control.ChoiceBox;
 import org.junit.Before;
 import org.junit.Test;
